+++ src/main/resources/db/changelog/changeset/V001.sql
@@ -4,6 +4,6 @@ create table if not exists request_filters (
     to_date timestamptz,
     host text,
     path text,
-    avg_headers integer,
-    avg_query_params integer
+    avg_headers decimal(5, 2),
+    avg_query_params decimal(5, 2)
 );
